renamed data folder to db, for clarity
